server/requirements.txt
flask 
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
pysnmp
paho-mqtt
Flask-MQTT
Werkzeug==2.3.7
